+++ src/Sulu/Component/Content/Metadata/Loader/XmlLoader.php
@@ -53,7 +53,7 @@ class XmlLoader extends XmlLegacyLoader
         $structure->cacheLifetime = $data['cacheLifetime'];
         $structure->controller = $data['controller'];
         $structure->internal = $data['internal'] === 'true';
-        $structure->zones = $data['zones'];
+        $structure->areas = $data['areas'];
         $structure->view = $data['view'];
         $structure->tags = $data['tags'];
         $structure->parameters = $data['params'];
@@ -200,7 +200,7 @@ class XmlLoader extends XmlLegacyLoader
                 'controller' => null,
                 'internal' => false,
                 'cacheLifetime' => null,
-                'zones' => [],
+                'areas' => [],
             ],
             $this->normalizeItem($data)
         );